+++ NEWS.md
@@ -2,6 +2,7 @@
 
 #### Updates
 
+* `ncurl()` and variants now accept request data supplied as a raw vector (thanks @thomasp85, #158).
 * `cv_reset()` now correctly resets the flag in the case a flag event has already been registered with `pipe_notify()`.
 * The previous `listen()` and `dial()` argument `error`, removed in v1.6.0, is now defunct.
 * The previous `serial_config()` argument `vec`, unutilised since v1.6.0, is removed.